+++ backend/redis/activity.go
@@ -3,9 +3,9 @@ package redis
 import (
 	"context"
 
+	"github.com/cschleiden/go-workflows/backend/task"
 	"github.com/cschleiden/go-workflows/internal/core"
 	"github.com/cschleiden/go-workflows/internal/history"
-	"github.com/cschleiden/go-workflows/internal/task"
 )
 
 func (rb *redisBackend) GetActivityTask(ctx context.Context) (*task.Activity, error) {
